PR126-19 #comment Separating upload functions and removing pipeline. Process should be run as individual steps to avoid duplicating downloads etc.
