+++ README.md
@@ -12,19 +12,7 @@ To insert files to the database, install PostGIS with the [`raster2pgsql`](https
 
 ---
 
-To download, resample, insert into database(s), and upload to S3 use the `pipeline` function. Make sure no rasters matching those you're processing
-have already been added to the database. There are no constraints in the tables to prevent duplicates.
-
-```python
-from worldpop import pipeline
-
-iso3_code = "USA"
-
-for year in range(2000, 2021):
-    pipeline(iso3_code, year)
-```
-
-There are also functions to run part of the pipeline:
+The pipeline to download, resample, insert into database(s), and upload to S3 use the functions listed below. Make sure no rasters matching those you're processing have already been added to the database. There are no constraints in the tables to prevent duplicates.
 
 - `build_urls` - given a country and year, return endpoints for all rasters
 - `download` - download raw data from World Pop
@@ -34,22 +22,26 @@ There are also functions to run part of the pipeline:
 - `insert_from_s3` - insert rasters to database that have already been uploaded to S3
 - `remove` - delete raster from current directory
 
-### Alternative Workflow
+### An example workflow
 
 ```python
 import os
 import worldpop as wp
 
-iso3_code = "LBR"
+iso3_code = "LBN"
 
 for year in range(2000, 2021):
     print(f"\nProcessing {year}: ", end="", flush=True)
     for i, url in enumerate(wp.build_urls(iso3_code, year)):
         print(f"{i+1}..", end="", flush=True)
         out_path = os.path.basename(url)
+        key = wp.s3_key(out_path)
+        if wp.exists_on_s3(key):
+            continue
         wp.download(url)
         wp.resample(out_path)
         wp.upload_to_s3(out_path)
         os.remove(out_path)
-    wp.insert_from_s3(iso3_code, year, tile_size=400, dev=True)
+# Before running any data base functions, check that the data hasn't already been added
+wp.insert_from_s3(iso3_code, year, tile_size=400, dev=True)
 ```
