+++ worldpop/__init__.py
@@ -1,4 +1,4 @@
-from .download import build_urls, download, upload_to_s3
+from .download import build_urls, download
 from .insert import insert, insert_from_s3
-from .pipeline import pipeline
+from .upload import exists_on_s3, s3_key, upload_to_s3
 from .utils import resample
